pyproject.toml
[project]
name = "airtable-pbgl"
version = "0.1.0"
description = "AirTable PBGL internal CLI and Textual TUI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "lxml>=6.0.2",
    "python-dotenv>=1.0",
    "requests>=2.32",
    "textual>=0.80",
    "typer>=0.15",
]

[project.scripts]
atp = "atp.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/atp"]

[dependency-groups]
dev = [
    "mypy",
    "pytest>=8",
    "pytest-cov",
    "ruff",
]

[tool.pytest.ini_options]
addopts = ["--import-mode=importlib"]
pythonpath = ["src"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.12"
strict = false
